move CloseButton to shared/CloseButton
