+++ src/components/SourceTabs.js
@@ -12,7 +12,7 @@ const { getFilename, getRawSourceURL, isPretty } = require("../utils/source");
 const { isEnabled } = require("devtools-config");
 const classnames = require("classnames");
 const actions = require("../actions");
-const CloseButton = require("./CloseButton");
+const CloseButton = require("./shared/CloseButton");
 const PaneToggleButton = require("./PaneToggleButton");
 const Svg = require("./utils/Svg");
 const Dropdown = React.createFactory(require("./Dropdown"));
