+++ src/components/shared/CloseButton.js
@@ -1,6 +1,6 @@
 const React = require("react");
 const { DOM: dom, PropTypes } = React;
-const Svg = require("./utils/Svg");
+const Svg = require("../utils/Svg");
 
 require("./CloseButton.css");
